Add support for capturing variables from pattern expressions.
This changes the expansion of pattern expressions when used in a labeled
sequence such that the remainder of the sequence is lexically within the
match arm so that the variable is in scope.

Literal expressions are also included for the sake of generating nicer
code by not generating RuleResult in a `match` that is itself
immediately matched upon.

Fixes #227